Move `GatheredImports` to `util_rules/import_analysis.py`
This does introduce a cycle when type hints are used. Not ideal, but these classes are directly related to import analysis so seem to belong here. I'll try to fix this cycle in followups.

# Rust tests and lints will be skipped. Delete if not intended.
[ci skip-rust]

# Building wheels and fs_util will be skipped. Delete if not intended.
[ci skip-build-wheels]
